More functions!
Defined the guessLetter and checkLetter functions, tinkered a little
with both guessLetter and guessWord functions, so that now instead of
being void, they return the guess that the user made. (Maybe that’s not
what we want to do? I’m actually not sure…) I realized that the
checkLetter function isn’t yet doing nearly all of what we need it to
do. Check the function itself for a lengthier comment about that.