tweaking a few functions in game.cpp
I was getting some errors with some of the functions so I took a look
at what was going wrong. There was a capitalization error with the
guessletter variable (not to be confused with the function,
guessLetter), and the findLetterPos function was being called with more
arguments than had been declared in the header file.