Whole lotta code!
makeTurn is now a bool, which returns whether or not the game has been
won. (Problem I just noticed: This will tell us if the game has been
won, but not who has won it. We’ll have to work on that.)
Also, there’s now a loop that will limit the game to six turns, and the
whole thing is in a loop that will ask the player if they want to play
again.